+++ src/main/java/com/github/temasaur/callstat/repository/SubscriberRepository.java
@@ -3,5 +3,5 @@ package com.github.temasaur.callstat.repository;
 import com.github.temasaur.callstat.models.Subscriber;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubscriberRepository extends CrudRepository<Subscriber, Integer> {
+public interface SubscriberRepository extends CrudRepository<Subscriber, Long> {
 }
